WIP: refactor: use enum to distinguish transaction signature types
