+++ crates/edr_eth/src/transaction/request/eip2930.rs
@@ -7,7 +7,7 @@ use revm_primitives::keccak256;
 use crate::{
     access_list::AccessListItem,
     signature::{self, SignatureError},
-    transaction::{self, fake_signature::make_fake_signature, TxKind},
+    transaction::{self, TxKind},
     utils::envelop_bytes,
     Address, Bytes, B256, U256,
 };
@@ -40,8 +40,6 @@ impl Eip2930 {
     ) -> Result<transaction::signed::Eip2930, SignatureError> {
         let hash = self.hash();
 
-        let signature = signature::Ecdsa::new(hash, secret_key)?;
-
         Ok(transaction::signed::Eip2930 {
             chain_id: self.chain_id,
             nonce: self.nonce,
@@ -51,18 +49,13 @@ impl Eip2930 {
             value: self.value,
             input: self.input,
             access_list: self.access_list.into(),
-            odd_y_parity: signature.odd_y_parity(),
-            r: signature.r,
-            s: signature.s,
+            signature: signature::Recoverable::rs_and_y_parity(hash, secret_key)?,
             hash: OnceLock::new(),
-            is_fake: false,
         })
     }
 
     /// Creates a fake signature for an impersonated account.
-    pub fn fake_sign(self, address: &Address) -> transaction::signed::Eip2930 {
-        let signature = make_fake_signature::<1>(address);
-
+    pub fn fake_sign(self, address: Address) -> transaction::signed::Eip2930 {
         transaction::signed::Eip2930 {
             chain_id: self.chain_id,
             nonce: self.nonce,
@@ -72,11 +65,8 @@ impl Eip2930 {
             value: self.value,
             input: self.input,
             access_list: self.access_list.into(),
-            odd_y_parity: signature.odd_y_parity(),
-            r: signature.r,
-            s: signature.s,
+            signature: signature::Recoverable::fake(address, 1, true),
             hash: OnceLock::new(),
-            is_fake: true,
         }
     }
 }
@@ -170,7 +160,7 @@ mod tests {
 
         let fake_sender: Address = "0xa5bc06d4548a3ac17d72b372ae1e416bf65b8ead".parse()?;
 
-        let signed = transaction.fake_sign(&fake_sender);
+        let signed = transaction.fake_sign(fake_sender);
 
         // Generated by Hardhat
         let expected_hash: B256 =
