+++ crates/edr_eth/src/transaction/request/legacy.rs
@@ -6,7 +6,7 @@ use revm_primitives::keccak256;
 
 use crate::{
     signature::{self, SignatureError},
-    transaction::{self, fake_signature::make_fake_signature, TxKind},
+    transaction::{self, TxKind},
     Address, Bytes, B256, U256,
 };
 
@@ -43,16 +43,13 @@ impl Legacy {
             kind: self.kind,
             value: self.value,
             input: self.input,
-            signature,
+            signature: signature.into(),
             hash: OnceLock::new(),
-            is_fake: false,
         })
     }
 
     /// Creates a fake signature for an impersonated account.
-    pub fn fake_sign(self, sender: &Address) -> transaction::signed::Legacy {
-        let signature = make_fake_signature::<0>(sender);
-
+    pub fn fake_sign(self, sender: Address) -> transaction::signed::Legacy {
         transaction::signed::Legacy {
             nonce: self.nonce,
             gas_price: self.gas_price,
@@ -60,9 +57,8 @@ impl Legacy {
             kind: self.kind,
             value: self.value,
             input: self.input,
-            signature,
+            signature: signature::Recoverable::fake(sender, 0, false),
             hash: OnceLock::new(),
-            is_fake: true,
         }
     }
 }
@@ -139,7 +135,7 @@ mod tests {
 
         let fake_sender: Address = "0xa5bc06d4548a3ac17d72b372ae1e416bf65b8ead".parse()?;
 
-        let signed = transaction.fake_sign(&fake_sender);
+        let signed = transaction.fake_sign(fake_sender);
 
         // Generated by Hardhat
         let expected_hash: B256 =
