+++ crates/edr_eth/src/transaction/signed/eip1559.rs
@@ -8,13 +8,13 @@ use super::kind_to_transact_to;
 use crate::{
     access_list::AccessList,
     signature::{self, SignatureError},
-    transaction::{self, fake_signature::recover_fake_signature, TxKind},
+    transaction::{self, TxKind},
     utils::envelop_bytes,
     Address, Bytes, B256, U256,
 };
 
 #[derive(Clone, Debug, Eq, RlpDecodable, RlpEncodable)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub struct Eip1559 {
     // The order of these fields determines de-/encoding order.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde::u64"))]
@@ -29,19 +29,13 @@ pub struct Eip1559 {
     pub value: U256,
     pub input: Bytes,
     pub access_list: AccessList,
-    pub odd_y_parity: bool,
-    pub r: U256,
-    pub s: U256,
+    #[cfg_attr(feature = "serde", serde(flatten))]
+    pub signature: signature::Recoverable,
     /// Cached transaction hash
     #[rlp(default)]
     #[rlp(skip)]
     #[cfg_attr(feature = "serde", serde(skip))]
     pub hash: OnceLock<B256>,
-    /// Whether the signature is from an impersonated account.
-    #[rlp(default)]
-    #[rlp(skip)]
-    #[cfg_attr(feature = "serde", serde(skip))]
-    pub is_fake: bool,
 }
 
 impl Eip1559 {
@@ -56,17 +50,8 @@ impl Eip1559 {
 
     /// Recovers the Ethereum address which was used to sign the transaction.
     pub fn recover(&self) -> Result<Address, SignatureError> {
-        let signature = signature::Ecdsa {
-            r: self.r,
-            s: self.s,
-            v: u64::from(self.odd_y_parity),
-        };
-
-        if self.is_fake {
-            return Ok(recover_fake_signature(&signature));
-        }
-
-        signature.recover(transaction::request::Eip1559::from(self).hash())
+        self.signature
+            .recover_address(transaction::request::Eip1559::from(self).hash())
     }
 
     /// Converts this transaction into a `TxEnv` struct.
@@ -101,9 +86,7 @@ impl PartialEq for Eip1559 {
             && self.value == other.value
             && self.input == other.input
             && self.access_list == other.access_list
-            && self.odd_y_parity == other.odd_y_parity
-            && self.r == other.r
-            && self.s == other.s
+            && self.signature == other.signature
     }
 }
