+++ crates/edr_eth/src/transaction/signed/eip4844.rs
@@ -7,13 +7,13 @@ use revm_primitives::{keccak256, TransactTo, TxEnv, GAS_PER_BLOB};
 use crate::{
     access_list::AccessList,
     signature::{self, SignatureError},
-    transaction::{self, fake_signature::recover_fake_signature},
+    transaction,
     utils::envelop_bytes,
     Address, Bytes, B256, U256,
 };
 
 #[derive(Clone, Debug, Eq, RlpDecodable, RlpEncodable)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub struct Eip4844 {
     // The order of these fields determines de-/encoding order.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde::u64"))]
@@ -30,19 +30,13 @@ pub struct Eip4844 {
     pub access_list: AccessList,
     pub max_fee_per_blob_gas: U256,
     pub blob_hashes: Vec<B256>,
-    pub odd_y_parity: bool,
-    pub r: U256,
-    pub s: U256,
+    #[cfg_attr(feature = "serde", serde(flatten))]
+    pub signature: signature::Recoverable,
     /// Cached transaction hash
     #[rlp(default)]
     #[rlp(skip)]
     #[cfg_attr(feature = "serde", serde(skip))]
     pub hash: OnceLock<B256>,
-    /// Whether the signed transaction is from an impersonated account.
-    #[rlp(default)]
-    #[rlp(skip)]
-    #[cfg_attr(feature = "serde", serde(skip))]
-    pub is_fake: bool,
 }
 
 impl Eip4844 {
@@ -61,17 +55,8 @@ impl Eip4844 {
 
     /// Recovers the Ethereum address which was used to sign the transaction.
     pub fn recover(&self) -> Result<Address, SignatureError> {
-        let signature = signature::Ecdsa {
-            r: self.r,
-            s: self.s,
-            v: u64::from(self.odd_y_parity),
-        };
-
-        if self.is_fake {
-            return Ok(recover_fake_signature(&signature));
-        }
-
-        signature.recover(transaction::request::Eip4844::from(self).hash())
+        self.signature
+            .recover_address(transaction::request::Eip4844::from(self).hash())
     }
 
     /// Converts this transaction into a `TxEnv` struct.
@@ -113,9 +98,7 @@ impl PartialEq for Eip4844 {
             && self.input == other.input
             && self.access_list == other.access_list
             && self.blob_hashes == other.blob_hashes
-            && self.odd_y_parity == other.odd_y_parity
-            && self.r == other.r
-            && self.s == other.s
+            && self.signature == other.signature
     }
 }
 
@@ -144,13 +127,18 @@ mod tests {
                 "0x01b0a4cdd5f55589f5c5b4d46c76704bb6ce95c0a8c09f77f197a57808dded28",
             )
             .unwrap()],
-            r: U256::from_str("0x8a83833ec07806485a4ded33f24f5cea4b8d4d24dc8f357e6d446bcdae5e58a7")
+            signature: signature::Recoverable::RsyParity {
+                r: U256::from_str(
+                    "0x8a83833ec07806485a4ded33f24f5cea4b8d4d24dc8f357e6d446bcdae5e58a7",
+                )
                 .unwrap(),
-            s: U256::from_str("0x68a2ba422a50cf84c0b5fcbda32ee142196910c97198ffd99035d920c2b557f8")
+                s: U256::from_str(
+                    "0x68a2ba422a50cf84c0b5fcbda32ee142196910c97198ffd99035d920c2b557f8",
+                )
                 .unwrap(),
-            odd_y_parity: false,
+                y_parity: false,
+            },
             hash: OnceLock::new(),
-            is_fake: false,
         }
     }
 
@@ -196,15 +184,16 @@ mod tests {
             blob_hashes: vec![b256!(
                 "01ae39c06daecb6a178655e3fab2e56bd61e81392027947529e4def3280c546e"
             )],
-            odd_y_parity: true,
-            r: U256::from_str(
-                "0xaeb099417be87077fe470104f6aa73e4e473a51a6c4be62607d10e8f13f9d082",
-            )?,
-            s: U256::from_str(
-                "0x390a4c98aaecf0cfc2b27e68bdcec511dd4136356197e5937ce186af5608690b",
-            )?,
+            signature: signature::Recoverable::RsyParity {
+                r: U256::from_str(
+                    "0xaeb099417be87077fe470104f6aa73e4e473a51a6c4be62607d10e8f13f9d082",
+                )?,
+                s: U256::from_str(
+                    "0x390a4c98aaecf0cfc2b27e68bdcec511dd4136356197e5937ce186af5608690b",
+                )?,
+                y_parity: true,
+            },
             hash: OnceLock::new(),
-            is_fake: false,
         };
 
         assert_eq!(transaction.recover()?, CALLER);
